Add matthew read in
Added code to read in the new version of the dataset. There are some variables that are no longer included in the dataset, and it's not clear what exact changes were made. Were they made manually?

We should condense this down to one method of reading in data later, perhaps with a branch and pull request.