Allow time effect to vary by metabolite
Conducted an "interaction analysis" with the random effects, basically allowing the time effect to have an overall intercept, but also to be nudged a little bit in one direction or another based upon the metabolite. Seems pretty cool! I tried nested random effects as well, but that didn't seem to make as much sense.

Closes #9