Switch to saving files as rds
When saving files as RData, you save both the name and the value of the object. This behavior seems to have caused issues multiples times when I am trying to read in the objects later, for example, to do posterior predictive draws. Therefore, I am switching to the rds format, which stores only the value of the object and allows you to assign it a new name when you read it in.