Improve estimation of doses
Instead of estimating doses per day per 80,000 people, which doesn't make a lot of sense, we have switched to estimating the mass load in each 30 minute block. We can go to doses from there. Also, when we are estimating the mass load by location, I split the flow of water evenly across the 3 locations, because the flow of water is split across there. I also reported it by 1,000 instead of the whole stadium, since there are different numbers of people in each section.