Modify how mass load per location is calculated
The flow rate wasn't being correctly applied to each locatino (i.e., divided by 3). So I fixed that, as well as changed the predictions by location to be mass load. I also made the plots dot plots, instead of trying to show the distributions. I think people will get confused that they are 95% confidence intervals and/or histograms of the distribution. It's a big amount of data reduction, but we can always hand over the more high-resolution results to people.