Update/verify the predicted number of doses.
The calculation of the doses/mass load over the entire stadium and time is complicated. You have to get the summaries, orders of groupings, and use of the knowledge about the proportion of the stadium covered by the location exactly right.

This commit represents the clearest and easiest to understand (as well as best-documented) execution of those calculations.