Estimate doses of parent compounds using metabolites
One of the reviewer issues was estimating doses using the raw amount of the substance in the wastewater. This quantity doesn't get at consumption very well. Therefore, we estimated the number of doses of cocaine, hydrocodone, and oxycodone using their metabolites (benzoylecgonine, norhydrocodone, and noroxycodone). We limited our graph of doses for the entire stadium throughout the entire game to these three compounds, because all of the other substances are parent compounds, not metabolites. Additionally, the number of doses has changed a great deal.